similar_words_finder_app_LLM/requirements.txt
langchain
streamlit
langchain-community
langchain-openai
tiktoken
python-dotenv
faiss-cpu